Rename source files for new classes.
