+++ src/keywords/types.h
@@ -68,6 +68,6 @@
 #include "keywords/vec3double.h"
 #include "keywords/vec3integer.h"
 #include "keywords/vec3nodevalue.h"
-#include "keywords/vector_id.h"
-#include "keywords/vector_is.h"
+#include "keywords/vector_intdouble.h"
+#include "keywords/vector_intstring.h"
 #include "keywords/windowfunction.h"
