+++ src/keywords/vector_intstring.cpp
@@ -1,6 +1,6 @@
 /*
-    *** Keyword - Vector of Integer/Double Tuples
-    *** src/keywords/vector_id.cpp
+    *** Keyword - Vector of Integer/String Tuples
+    *** src/keywords/vector_intstring.cpp
     Copyright T. Youngs 2012-2020
 
     This file is part of Dissolve.
@@ -19,36 +19,36 @@
     along with Dissolve.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "keywords/vector_id.h"
+#include "keywords/vector_intstring.h"
 #include "base/lineparser.h"
 #include "classes/coredata.h"
 
-IntegerDoubleVectorKeyword::IntegerDoubleVectorKeyword::IntegerDoubleVectorKeyword(IntegerDoubleVectorKeywordData &data,
+IntegerStringVectorKeyword::IntegerStringVectorKeyword::IntegerStringVectorKeyword(IntegerStringVectorKeywordData &data,
                                                                                    int nRequiredIntegers, int nRequiredValues)
-    : KeywordData<IntegerDoubleVectorKeywordData &>(KeywordBase::VectorIntegerDoubleData, data),
+    : KeywordData<IntegerStringVectorKeywordData &>(KeywordBase::VectorIntegerStringData, data),
       nRequiredIntegers_(nRequiredIntegers), nRequiredValues_(nRequiredValues)
 {
 }
 
-IntegerDoubleVectorKeyword::~IntegerDoubleVectorKeyword() {}
+IntegerStringVectorKeyword::~IntegerStringVectorKeyword() {}
 
 // Return minimum number of arguments accepted
-int IntegerDoubleVectorKeyword::minArguments() const
+int IntegerStringVectorKeyword::minArguments() const
 {
     return (nRequiredIntegers_ + (nRequiredValues_ == -1 ? 1 : nRequiredValues_));
 }
 
 // Return maximum number of arguments accepted
-int IntegerDoubleVectorKeyword::maxArguments() const
+int IntegerStringVectorKeyword::maxArguments() const
 {
     return (nRequiredIntegers_ + (nRequiredValues_ == -1 ? 99 : nRequiredValues_));
 }
 
 // Parse arguments from supplied LineParser, starting at given argument offset
-bool IntegerDoubleVectorKeyword::read(LineParser &parser, int startArg, CoreData &coreData)
+bool IntegerStringVectorKeyword::read(LineParser &parser, int startArg, CoreData &coreData)
 {
     std::vector<int> i;
-    std::vector<double> d;
+    std::vector<std::string> s;
 
     auto argIndex = startArg;
 
@@ -63,8 +63,8 @@ bool IntegerDoubleVectorKeyword::read(LineParser &parser, int startArg, CoreData
 
     // Read values and check number
     for (auto n = argIndex; n < parser.nArgs(); ++n)
-        d.push_back(parser.argd(n));
-    if (d.size() < (nRequiredValues_ == -1 ? 1 : nRequiredValues_))
+        s.emplace_back(parser.argsv(n));
+    if (s.size() < (nRequiredValues_ == -1 ? 1 : nRequiredValues_))
     {
         if (nRequiredValues_ == -1)
             return Messenger::error("Keyword {} expects at least one double value in addition.\n", name());
@@ -74,7 +74,7 @@ bool IntegerDoubleVectorKeyword::read(LineParser &parser, int startArg, CoreData
     }
 
     // Add tuple to vector
-    data_.emplace_back(i, d);
+    data_.emplace_back(i, s);
 
     hasBeenSet();
 
@@ -82,15 +82,15 @@ bool IntegerDoubleVectorKeyword::read(LineParser &parser, int startArg, CoreData
 }
 
 // Write keyword data to specified LineParser
-bool IntegerDoubleVectorKeyword::write(LineParser &parser, std::string_view keywordName, std::string_view prefix)
+bool IntegerStringVectorKeyword::write(LineParser &parser, std::string_view keywordName, std::string_view prefix)
 {
-    for (const auto &idData : data_)
+    for (const auto &d : data_)
     {
         std::string line;
-        for (const auto &i : std::get<0>(idData))
+        for (const auto &i : std::get<0>(d))
             line += fmt::format("  {}", i);
-        for (const auto &d : std::get<1>(idData))
-            line += fmt::format("  '{}'", d);
+        for (const auto &s : std::get<1>(d))
+            line += fmt::format("  '{}'", s);
 
         if (!parser.writeLineF("{}{}{}\n", prefix, keywordName, line))
             return false;
